Split class diagram transaction applying process
